Compiles in Visual Studio.

+++ hgemmtest.cpp
@@ -322,8 +322,8 @@ int main() {
         auto sum_error = 0.0f;
 
         for (int i=0; i<4; i++) {
-            cl::NDRange local_sgemm = {32 * mdimc, ndimc, 1};
-            cl::NDRange size_sgemm = {32 * m / 16 * mdimc / mwg, n / 16 * ndimc / nwg, size_t(batch_size)};
+            cl::NDRange local_sgemm = {(size_t)32 * mdimc, (size_t)ndimc, 1};
+            cl::NDRange size_sgemm = {(size_t)32 * m / 16 * mdimc / mwg, (size_t)n / 16 * ndimc / nwg, size_t(batch_size)};
 
             queue.enqueueNDRangeKernel(kernel, cl::NullRange,
                                        size_sgemm, local_sgemm,
